+++ xmtp_mls/src/groups/device_sync.rs
@@ -1,9 +1,8 @@
-use super::group_metadata::ConversationType;
 use super::{GroupError, MlsGroup};
 use crate::configuration::NS_IN_HOUR;
 use crate::groups::scoped_client::LocalScopedGroupClient;
 use crate::retry::{RetryBuilder, RetryableError};
-use crate::storage::group::GroupQueryArgs;
+use crate::storage::group::{ConversationType, GroupQueryArgs};
 use crate::storage::group_message::MsgQueryArgs;
 use crate::storage::DbConnection;
 use crate::subscriptions::{LocalEvents, StreamMessages, SubscribeError, SyncMessage};
@@ -228,6 +227,10 @@ where
      * Will auto-send a sync request if sync group is created.
      */
     pub async fn sync_init(&self, provider: &XmtpOpenMlsProvider) -> Result<(), DeviceSyncError> {
+        tracing::info!(
+            "Initializing device sync... url: {:?}",
+            self.history_sync_url
+        );
         if self.get_sync_group().is_err() {
             self.ensure_sync_group(provider).await?;
 
@@ -236,6 +239,7 @@ where
             self.send_sync_request(provider, DeviceSyncKind::MessageHistory)
                 .await?;
         }
+        tracing::info!("Device sync initialized.");
 
         Ok(())
     }
@@ -261,6 +265,7 @@ where
         provider: &XmtpOpenMlsProvider,
         kind: DeviceSyncKind,
     ) -> Result<DeviceSyncRequestProto, DeviceSyncError> {
+        tracing::info!("Sending a sync request for {kind:?}");
         let request = DeviceSyncRequest::new(kind);
 
         // find the sync group