+++ xmtp_mls/src/groups/group_metadata.rs
@@ -7,6 +7,8 @@ use xmtp_proto::xmtp::mls::message_contents::{
     GroupMetadataV1 as GroupMetadataProto, Inbox as InboxProto,
 };
 
+use crate::storage::group::ConversationType;
+
 #[derive(Debug, Error)]
 pub enum GroupMetadataError {
     #[error("serialization: {0}")]
@@ -104,12 +106,6 @@ impl TryFrom<&Extensions> for GroupMetadata {
  * *DM*: A conversation between 2 members with simplified permissions
  * *Sync*: A conversation between all the devices of a single member with simplified permissions
  */
-#[derive(Debug, Copy, Clone, PartialEq)]
-pub enum ConversationType {
-    Group,
-    Dm,
-    Sync,
-}
 
 impl From<ConversationType> for ConversationTypeProto {
     fn from(value: ConversationType) -> Self {